apps/requirements.txt
dash
dash-bootstrap-components
databricks-sdk
python-dotenv
databricks-sdk==0.52.0
openai==1.55.3
flask==3.0.3
logging